requirements.txt
# This file contains the build requirements for ZenDNN_PACE.
# It is used to install the necessary packages for building the project.
-r build_requirements.txt
torchvision==0.22.0+cpu     # required by lm_eval and transformers
black==25.1.0
numpy>=2.0.2
hypothesis>=6.113.0
huggingface-hub[hf_xet]==0.31.4
transformers==4.51.3
safetensors>=0.5.2
flake8>=7.1.1
expecttest>=0.2.1
psutil>=7.0.0
matplotlib>=3.9.4
datasets>=3.3.2
lm_eval==0.4.7
langdetect==1.0.9
immutabledict==4.2.1
protobuf==6.31.0
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
httpx==0.28.1
backports.strenum (>=1.3.1,<2.0) ; python_version < '3.11'